refactor(shared): split types.ts into domain modules with barrel re-export
Audit finding F3. The 1080-line god file is replaced by six domain files
under src/<domain>/types.ts plus a thin barrel that preserves all existing
imports unchanged. ElectronAPI is moved out of shared entirely into
apps/dm-tool/electron/ipc/types.ts (it had no consumers outside dm-tool).
