+++ apps/dm-tool/electron/preload.ts
@@ -7,45 +7,45 @@
 // src/vite-env.d.ts which declares `window.electronAPI` for TypeScript.
 
 import { contextBridge, ipcRenderer } from 'electron';
+import type { ElectronAPI } from './ipc/types.js';
 import type {
   ActorSpellcasting,
   ActorUpdate,
   AonPreviewData,
   AurusTeam,
-  CombatantInitiativeEvent,
-  Encounter,
-  LootItem,
-  PartyMember,
-  PushEncounterResult,
   Book,
   BookClassifyProgress,
   BookScanResult,
   ChatChunk,
   ChatMessage,
   ChatModel,
+  CombatantInitiativeEvent,
   CompendiumPackSummary,
   ConfigPaths,
-  ElectronAPI,
+  Encounter,
   Facets,
   FinalizeIngestArgs,
   GlobePin,
-  MissionData,
   ItemBrowserDetail,
   ItemBrowserRow,
   ItemFacets,
   ItemSearchParams,
+  LootItem,
   MapDetail,
   MapSummary,
+  MissionData,
   MonsterDetail,
   MonsterFacets,
   MonsterSearchParams,
   MonsterSummary,
   PartyInventoryItem,
+  PartyMember,
   PickPathArgs,
+  PushEncounterResult,
   SearchParams,
   TaggerProgress,
-  TaggerRunArgs,
   TaggerResult,
+  TaggerRunArgs,
 } from '@foundry-toolkit/shared/types';
 
 const api: ElectronAPI = {
