+++ apps/dm-tool/src/lib/api.ts
@@ -2,6 +2,6 @@
 // imports from a single module instead of sprinkling `window.electronAPI`
 // calls, which makes it easier to mock in tests later.
 
-import type { ElectronAPI } from '@foundry-toolkit/shared/types';
+import type { ElectronAPI } from '../../electron/ipc/types.js';
 
 export const api: ElectronAPI = window.electronAPI;
