+++ packages/shared/src/types.ts
@@ -1,401 +1,19 @@
-// Shared types between the Electron main process and the React renderer.
+// Barrel re-export — keeps all existing `import ... from '@foundry-toolkit/shared'`
+// and `import ... from '@foundry-toolkit/shared/types'` working without changes.
 //
-// These mirror the shape of rows in the map-tagger SQLite index and of the
-// JSON sidecars written next to each map. Keep this file dependency-free so
-// both sides of the contextBridge can import it without pulling in runtime
-// modules they don't have access to.
+// New types belong in the appropriate domain file under src/<domain>/types.ts.
+// This barrel is preserved for one minor-version cycle; see audit finding F3.
 
-/** Controlled vocabulary — kept loose on the TS side since the DB is the
- *  source of truth and we don't want to break the app if the Python side
- *  adds a new value. */
-export type InteriorExterior = 'interior' | 'exterior' | 'mixed' | 'unknown';
-export type TimeOfDay = 'day' | 'dusk' | 'night' | 'dawn' | 'unknown';
-export type GridVisible = 'gridded' | 'gridless' | 'unknown';
-
-/** A lightweight row used by the browser grid. This is what `searchMaps`
- *  returns — enough to render a thumbnail, title, and a few chips without
- *  the full sidecar payload. */
-export interface MapSummary {
-  fileName: string;
-  title: string;
-  description: string;
-  interiorExterior: InteriorExterior | null;
-  timeOfDay: TimeOfDay | null;
-  gridVisible: GridVisible | null;
-  gridCells: string | null;
-  approxPartyScale: string | null;
-}
-
-/** The full sidecar contents for a single map, used by the detail pane. */
-export interface MapDetail extends MapSummary {
-  fileHashSha256: string;
-  phash: string;
-  widthPx: number;
-  heightPx: number;
-  biomes: string[];
-  locationTypes: string[];
-  mood: string[];
-  features: string[];
-  encounterHooks: string[];
-  /** Additional encounter hooks generated client-side via the Anthropic
-   *  API and persisted to a dm-tool-owned override file (see
-   *  electron/hooks-store.ts). The DB is read-only, so these can't live
-   *  in the sidecar JSON. Newest first — UI prepends to this list when
-   *  the user clicks the refresh button. */
-  additionalEncounterHooks: string[];
-  taggedAt: string; // ISO 8601
-  model: string;
-}
-
-export interface SearchParams {
-  keywords?: string;
-  biomes?: string[];
-  locationTypes?: string[];
-  mood?: string[];
-  features?: string[];
-  interiorExterior?: InteriorExterior;
-  timeOfDay?: TimeOfDay;
-  gridVisible?: GridVisible;
-  limit?: number;
-}
-
-/** Distinct tag values returned by `listFacets` so the filter panel can
- *  show checkboxes without hardcoding the enum lists. */
-export interface Facets {
-  biomes: string[];
-  locationTypes: string[];
-  moods: string[];
-  features: string[];
-}
-
-// ---------------------------------------------------------------------------
-// Book catalog + reader
-// ---------------------------------------------------------------------------
-
-/** One row in the `books` table. Matches the SQLite schema in
- *  electron/book-db.ts one-to-one except for naming (snake_case in SQL,
- *  camelCase here) and the omitted `path` field — the renderer never sees
- *  the absolute path, it accesses PDFs via `book-file://files/<id>` URLs
- *  served by the main process. */
-export interface Book {
-  id: number;
-  title: string;
-  category: string;
-  subcategory: string | null;
-  ruleset: 'legacy' | 'remastered' | null;
-  pageCount: number | null;
-  fileSize: number;
-  /** True once phase-2 ingest has run — we have a cached cover PNG and a
-   *  real page count. Covers are fetched by the renderer via
-   *  `book-file://covers/<id>` URLs; if this is false, the UI should show
-   *  a placeholder instead of a broken image. */
-  ingested: boolean;
-  // AI classification (null until classified)
-  aiSystem: string | null;
-  aiCategory: string | null;
-  aiSubcategory: string | null;
-  aiTitle: string | null;
-  aiPublisher: string | null;
-  classified: boolean;
-}
-
-export interface BookClassification {
-  system: string;
-  category: string;
-  subcategory: string | null;
-  title: string;
-  publisher: string | null;
-}
-
-export interface BookClassifyProgress {
-  type: 'progress' | 'done' | 'error';
-  bookId?: number;
-  bookTitle?: string;
-  current?: number;
-  total?: number;
-  error?: string;
-}
-
-/** Result of a phase-1 scan. Summary counts only — if the renderer needs
- *  the new data it calls `listBooks()` after the scan resolves. */
-export interface BookScanResult {
-  added: number;
-  updated: number;
-  removed: number;
-  total: number;
-}
-
-/** Renderer-to-main IPC payload for the phase-2 ingest finalize step. The
- *  renderer does the PDF rendering (so we don't have to build node-canvas
- *  on Windows) and ships the cover PNG bytes back as a plain Uint8Array
- *  via structured clone. */
-export interface FinalizeIngestArgs {
-  id: number;
-  pageCount: number;
-  /** Raw bytes of a 300 px-wide PNG. Main writes this to
-   *  `<userData>/book-covers/<id>.png`. */
-  coverPngBytes: Uint8Array;
-}
-
-// ---------------------------------------------------------------------------
-// Item browser
-// ---------------------------------------------------------------------------
-
-export type ItemSortField = 'name' | 'level' | 'price';
-export type SortDirection = 'asc' | 'desc';
-
-export interface ItemSearchParams {
-  keywords?: string;
-  levelMin?: number;
-  levelMax?: number;
-  rarities?: string[];
-  isMagical?: boolean | null;
-  usageCategories?: string[];
-  traits?: string[];
-  sources?: string[];
-  sortBy?: ItemSortField;
-  sortDir?: SortDirection;
-  limit?: number;
-}
-
-/** Lightweight row for the item table — no description to keep IPC payloads
- *  small when returning hundreds of results. */
-export interface ItemBrowserRow {
-  id: string;
-  name: string;
-  level: number | null;
-  traits: string[];
-  rarity: string;
-  price: string | null;
-  bulk: string | null;
-  usage: string | null;
-  isMagical: boolean;
-  hasVariants: boolean;
-  /** true = ORC/remastered, false = OGL/legacy, null = unknown. */
-  isRemastered: boolean | null;
-  /** Renderer-ready URL for the item's Foundry icon (already routed through
-   *  the asset proxy or monster-file:// protocol by the IPC layer). Null
-   *  when the icon is missing or is a default-icon placeholder. */
-  img?: string | null;
-}
-
-export interface ItemVariant {
-  type: string;
-  level: number | null;
-  price: string | null;
-}
-
-/** Full item detail including description and parsed variants. */
-export interface ItemBrowserDetail extends ItemBrowserRow {
-  description: string;
-  source: string | null;
-  aonUrl: string | null;
-  variants: ItemVariant[];
-  hasActivation: boolean;
-  /** Foundry document type slug: "weapon" | "armor" | "consumable" | "equipment" | "shield" | etc. */
-  itemType: string;
-}
-
-/** Distinct filter values for the item filter panel. */
-export interface ItemFacets {
-  traits: string[];
-  sources: string[];
-  usageCategories: string[];
-}
-
-/** Renderer-facing compendium pack descriptor. Mirrors the subset of
- *  foundry-mcp's `CompendiumPack` that the Settings → Monsters dialog
- *  needs — enough to render each pack as a labeled checkbox. */
-export interface CompendiumPackSummary {
-  /** Foundry pack id, e.g. `pf2e.pathfinder-bestiary`. Canonical key
-   *  for the monster-pack override setting. */
-  id: string;
-  /** Human-readable label from Foundry's pack metadata, e.g. "Pathfinder
-   *  Bestiary". Shown as the checkbox label. */
-  label: string;
-  /** 'Actor' for bestiaries/NPCs, 'Item' for equipment/feats/etc. */
-  type: string;
-  /** Game system name. Usually `'pf2e'` for the packs this dialog lists. */
-  system?: string;
-}
-
-// ---------------------------------------------------------------------------
-// Chat
-// ---------------------------------------------------------------------------
-
-export type ChatModel = 'claude-sonnet-4-6' | 'claude-haiku-4-5-20251001' | 'claude-opus-4-6';
-
-/** Single source of truth for the Claude model used by @foundry-toolkit/ai agents
- *  and the renderer's picker default. Downstream constants files re-export
- *  this under their own names (DEFAULT_MODEL, DEFAULT_CHAT_MODEL). */
-export const DEFAULT_CHAT_MODEL: ChatModel = 'claude-sonnet-4-6';
-
-export interface ChatMessage {
-  role: 'user' | 'assistant';
-  content: string;
-}
-
-export interface ChatChunk {
-  type: 'delta' | 'done' | 'error' | 'tool-status';
-  text?: string;
-  error?: string;
-}
-
-export interface AonCreaturePreview {
-  type: 'creature';
-  name: string;
-  level: number;
-  hp: number;
-  ac: number;
-  fortitude: number;
-  reflex: number;
-  will: number;
-  perception: number;
-  speed: string;
-  size: string;
-  traits: string[];
-  abilities: string[];
-  immunities: string[];
-  weaknesses: string;
-  rarity: string;
-  summary: string;
-  strength: number;
-  dexterity: number;
-  constitution: number;
-  intelligence: number;
-  wisdom: number;
-  charisma: number;
-  /** Raw stat block text — everything after the first `---` separator. */
-  statBlock: string;
-}
-
-export interface AonGenericPreview {
-  type: 'generic';
-  name: string;
-  category: string;
-  text: string;
-}
-
-export type AonPreviewData = AonCreaturePreview | AonGenericPreview;
-
-// ---------------------------------------------------------------------------
-// Monster browser
-// ---------------------------------------------------------------------------
-
-export interface MonsterSearchParams {
-  keywords?: string;
-  levels?: [number, number];
-  rarities?: string[];
-  sizes?: string[];
-  creatureTypes?: string[];
-  traits?: string[];
-  sources?: string[];
-  hpMin?: number;
-  hpMax?: number;
-  acMin?: number;
-  acMax?: number;
-  fortMin?: number;
-  refMin?: number;
-  willMin?: number;
-  sortBy?: 'name' | 'level' | 'hp' | 'ac';
-  sortDir?: 'asc' | 'desc';
-  limit?: number;
-}
-
-export interface MonsterSummary {
-  name: string;
-  level: number;
-  hp: number;
-  ac: number;
-  fort: number;
-  ref: number;
-  will: number;
-  rarity: string;
-  size: string;
-  creatureType: string;
-  traits: string[];
-  source: string;
-  aonUrl: string;
-}
-
-export interface MonsterSpellInfo {
-  name: string;
-  /** 0 = cantrip, 1–10 = spell rank */
-  rank: number;
-  /** For innate spells — undefined means unlimited/at-will */
-  usesPerDay?: number;
-  /** PF2e action cost: "1", "2", "3", "reaction", "free", or empty */
-  castTime: string;
-  range: string;
-  area: string;
-  target: string;
-  traits: string[];
-  /** Cleaned plain-text description (Foundry markup stripped) */
-  description: string;
-}
-
-export interface MonsterSpellRank {
-  rank: number;
-  spells: MonsterSpellInfo[];
-}
-
-export interface MonsterSpellGroup {
-  entryName: string;
-  tradition: string;
-  castingType: string;
-  dc?: number;
-  attack?: number;
-  ranks: MonsterSpellRank[];
-}
-
-export interface MonsterDetail {
-  name: string;
-  level: number;
-  source: string;
-  rarity: string;
-  size: string;
-  traits: string[];
-  hp: number;
-  ac: number;
-  fort: number;
-  ref: number;
-  will: number;
-  perception: number;
-  skills: string;
-  str: number;
-  dex: number;
-  con: number;
-  int: number;
-  wis: number;
-  cha: number;
-  speed: string;
-  immunities: string;
-  weaknesses: string;
-  resistances: string;
-  melee: string;
-  ranged: string;
-  abilities: string;
-  /** Structured spell groups. Empty array when the monster has no spells. */
-  spells: MonsterSpellGroup[];
-  description: string;
-  aonUrl: string;
-  /** Relative path to portrait art image, or null if unavailable. */
-  imageUrl: string | null;
-  /** Relative path to token image, or null if unavailable. */
-  tokenUrl: string | null;
-}
-
-export interface MonsterFacets {
-  rarities: string[];
-  sizes: string[];
-  creatureTypes: string[];
-  traits: string[];
-  sources: string[];
-  levelRange: [number, number];
-}
+export * from './maps/types.js';
+export * from './books/types.js';
+export * from './compendium/types.js';
+export * from './chat/types.js';
+export * from './party/types.js';
+export * from './combat/types.js';
 
 // ---------------------------------------------------------------------------
 // Config (exposed to renderer for Settings UI / first-run setup)
+// Stragglers: only consumed by dm-tool's ElectronAPI and Settings UI.
 // ---------------------------------------------------------------------------
 
 /** All config paths surfaced to the renderer. Optional fields use "" when
@@ -419,662 +37,4 @@ export interface PickPathArgs {
   mode: 'directory' | 'file';
   title?: string;
   filters?: { name: string; extensions: string[] }[];
-}
-
-// ---------------------------------------------------------------------------
-// Map tagger
-// ---------------------------------------------------------------------------
-
-export interface TaggerRunArgs {
-  sourcePath: string;
-  apiKey: string;
-  limit: number;
-  concurrency?: number;
-}
-
-export interface TaggerProgress {
-  type: 'stdout' | 'stderr';
-  line: string;
-}
-
-export interface TaggerResult {
-  exitCode: number | null;
-  signal: string | null;
-}
-
-/** The IPC surface exposed to the renderer via contextBridge. Every
- *  function here must have a corresponding handler registered in ipc.ts
- *  and a corresponding type declaration on `window.electronAPI` in the
- *  renderer's global types. */
-// --- Globe pins --------------------------------------------------------------
-
-export type GlobePinKind = 'note' | 'mission';
-
-export interface GlobePin {
-  id: string;
-  lng: number;
-  lat: number;
-  label: string;
-  /** game-icons.net icon name (e.g. "crossed-swords"). Empty string = default dot. */
-  icon: string;
-  /** CSS hex fill color for the pin's circle (e.g. "#e03c31"). Empty/absent = default golden. */
-  iconColor?: string;
-  /** Zoom level at which the pin was placed. Icons shrink when zoomed out past this. */
-  zoom: number;
-  /** Relative path to the Obsidian note within the vault (e.g. "Golarion/My Pin a1b2c3d4.md"). Empty = no note yet. */
-  note: string;
-  /** Pin kind: generic note (opens Obsidian on dbl-click) or mission (opens in-universe briefing). */
-  kind: GlobePinKind;
-  /** Pre-parsed mission data, present only on mission pins in the exported
-   *  data.json consumed by the player-map. DB-stored pins never carry
-   *  this — missions are re-parsed on demand on the DM side. */
-  mission?: MissionData;
-}
-
-/** Shape of the data.json file the DM tool writes out and the player-map
- *  consumes as its static data source. */
-export interface ExportData {
-  exportedAt: string;
-  pins: GlobePin[];
-}
-
-// --- Mission briefing data (parsed from Obsidian frontmatter) ----------------
-
-export type MissionThreatLevel = 'Trivial' | 'Low' | 'Moderate' | 'Severe' | 'Extreme';
-export type MissionStatus = 'Available' | 'Assigned' | 'Active' | 'Completed' | 'Failed';
-
-export interface MissionObjective {
-  id: string;
-  text: string;
-  /** Primary objectives are mandatory; secondary are optional.
-   *  Frontmatter accepts either `primary: true` or `required: true`. */
-  isPrimary: boolean;
-  completed: boolean;
-}
-
-export interface MissionThreat {
-  id: string;
-  name: string;
-  /** Numeric level, or a string like "—" when not applicable (e.g. for
-   *  environmental hazards). Stored verbatim — the UI just displays it. */
-  level: number | string;
-  type?: string;
-}
-
-export interface MissionReward {
-  gold?: number;
-  xp?: number;
-  items?: string[];
-}
-
-export interface MissionData {
-  name: string;
-  threatLevel: MissionThreatLevel;
-  status: MissionStatus;
-  recommendedLevel: string;
-  estimatedSessions: string;
-  location: string;
-  questGiver: { name: string; title: string };
-  briefing: string[];
-  objectives: MissionObjective[];
-  threats: MissionThreat[];
-  rewards: MissionReward;
-  dmNotes: string;
-  datePosted: string;
-  sourceBook?: string;
-  /** Organizational branch issuing the posting (e.g. a faction arm). */
-  arm?: string;
-  /** Free-form description of the party/parties handling the mission. */
-  assignedTo?: string;
-  /** The target/prize of the mission — typically a named artifact. May
-   *  contain Obsidian-style `[[wikilinks]]`. */
-  artifact?: string;
-}
-
-export interface ElectronAPI {
-  // -----------------------------------------------------------------------
-  // Secure storage (OS keychain-backed via Electron safeStorage)
-  // -----------------------------------------------------------------------
-  secureStore(key: string, value: string): Promise<void>;
-  secureLoad(key: string): Promise<string>;
-  secureDelete(key: string): Promise<void>;
-
-  // -----------------------------------------------------------------------
-  // App mode + config
-  // -----------------------------------------------------------------------
-
-  /** Returns "setup" on first run (no config.json found), "normal" otherwise. */
-  getAppMode(): Promise<'normal' | 'setup'>;
-  /** Current config paths for display in the Settings UI. */
-  getConfig(): Promise<ConfigPaths>;
-  /** Open a native folder or file picker dialog. */
-  pickPath(args: PickPathArgs): Promise<string | null>;
-  /** Write config.json to userData and restart the app. */
-  saveConfigAndRestart(paths: ConfigPaths): Promise<void>;
-
-  searchMaps(params: SearchParams): Promise<MapSummary[]>;
-  getMapDetail(fileName: string): Promise<MapDetail | null>;
-  getFacets(): Promise<Facets>;
-  getLibraryPath(): Promise<string>;
-  openInExplorer(fileName: string): Promise<void>;
-  /** Update the native title bar overlay height at runtime so the OS
-   *  min/max/close button strip stays matched to the React header when
-   *  the user changes the UI scale in settings. */
-  setTitleBarOverlayHeight(height: number): Promise<void>;
-  /** Generate fresh encounter hooks for a map via the Anthropic API and
-   *  append them to the dm-tool override store. Returns the FULL list of
-   *  additional hooks (newest first) so the renderer can replace its
-   *  local state in one shot. The API key is passed in by the renderer
-   *  rather than read from disk in main — the renderer owns persistence
-   *  via localStorage and we want to avoid duplicating that. */
-  regenerateEncounterHooks(args: { fileName: string; apiKey: string }): Promise<string[]>;
-
-  /** Open a URL in the user's default browser. Only accepts http/https. */
-  openExternal(url: string): Promise<void>;
-  /** Fetch AoN preview data for a hover card. */
-  aonPreview(urlPath: string): Promise<AonPreviewData | null>;
-
-  // -----------------------------------------------------------------------
-  // Chat
-  // -----------------------------------------------------------------------
-
-  /** Send a chat message and begin streaming the assistant response.
-   *  Resolves when the stream completes. Text chunks arrive via
-   *  onChatChunk before the promise settles. */
-  chatSend(args: {
-    messages: ChatMessage[];
-    apiKey: string;
-    model?: ChatModel;
-    toolContext?: string;
-    rulesMode?: boolean;
-  }): Promise<void>;
-  /** Extract the visible text content from an embedded tool iframe by its base URL. */
-  getToolPageContent(toolUrl: string): Promise<string>;
-  /** Subscribe to chat stream chunks. Returns an unsubscribe function.
-   *  Same push-event pattern as onTaggerProgress. */
-  onChatChunk(callback: (chunk: ChatChunk) => void): () => void;
-
-  // -----------------------------------------------------------------------
-  // Book catalog + reader
-  // -----------------------------------------------------------------------
-
-  /** Walk the configured books root and reconcile the `books` table with
-   *  what's on disk. Cheap — file metadata only, no PDF parsing. Runs
-   *  automatically at startup; the UI also exposes a "Rescan" button for
-   *  manual refresh after the user adds/removes PDFs. */
-  booksScan(): Promise<BookScanResult>;
-  /** All rows in the catalog, sorted by category → subcategory → title. */
-  booksList(): Promise<Book[]>;
-  /** Fetch a single book by id. Returns null if the id is unknown (e.g.
-   *  the renderer had a stale list and the row was removed by a rescan). */
-  booksGet(id: number): Promise<Book | null>;
-  /** Finalize phase-2 ingest: main writes the cover PNG to
-   *  `<userData>/book-covers/<id>.png`, updates the page_count +
-   *  ingested_at columns, and returns the fresh Book row. */
-  booksFinalizeIngest(args: FinalizeIngestArgs): Promise<Book>;
-  /** The `book-file://files/<id>` URL the renderer should hand to pdfjs.
-   *  The main process resolves the id back to the absolute path at fetch
-   *  time, so the renderer never sees the real filesystem path — keeps
-   *  the renderer out of the file tree entirely. */
-  booksGetFileUrl(id: number): Promise<string>;
-  /** The `book-file://covers/<id>` URL the renderer should hand to an
-   *  <img> tag. Returns the URL even if the cover doesn't exist yet — the
-   *  <img> tag's onError handler will fall back to a placeholder, and
-   *  once ingest completes the URL starts resolving. */
-  booksGetCoverUrl(id: number): Promise<string>;
-
-  /** Update AI metadata fields for a single book (manual reclassification). */
-  booksUpdateMeta(args: {
-    id: number;
-    fields: { aiSystem?: string; aiCategory?: string; aiSubcategory?: string | null; aiPublisher?: string | null };
-  }): Promise<Book | null>;
-
-  /** Classify all unclassified books (or all if reclassify=true) using AI.
-   *  Progress streams via onBookClassifyProgress. */
-  booksClassify(args: { apiKey: string; reclassify?: boolean }): Promise<void>;
-  /** Cancel an in-progress classification run. */
-  booksClassifyCancel(): Promise<void>;
-  /** Subscribe to classification progress events. Returns unsubscribe fn. */
-  onBookClassifyProgress(callback: (p: BookClassifyProgress) => void): () => void;
-
-  // -----------------------------------------------------------------------
-  // Map tagger (ingest new maps)
-  // -----------------------------------------------------------------------
-
-  /** Whether the map-tagger binary is configured and available. If false,
-   *  the Add Maps UI should be hidden — the rest of the app still works
-   *  against the pre-tagged library. */
-  taggerAvailable(): Promise<boolean>;
-  /** Open a folder picker and return the selected path, or null if
-   *  cancelled. */
-  taggerPickSource(): Promise<string | null>;
-  /** Spawn the tagger in --preview mode to get a cost estimate without
-   *  calling the API. Progress lines stream via onTaggerProgress. */
-  taggerPreview(args: TaggerRunArgs): Promise<TaggerResult>;
-  /** Spawn the tagger for real ingest. Progress lines stream via
-   *  onTaggerProgress. Resolves when the process exits. */
-  taggerIngest(args: TaggerRunArgs): Promise<TaggerResult>;
-  /** Kill a running tagger process. Returns true if one was running. */
-  taggerCancel(): Promise<boolean>;
-  /** Returns true if a tagger subprocess is currently running. */
-  taggerIsRunning(): Promise<boolean>;
-  /** Subscribe to tagger progress events (stdout/stderr lines). Returns
-   *  an unsubscribe function. */
-  onTaggerProgress(callback: (p: TaggerProgress) => void): () => void;
-
-  // -----------------------------------------------------------------------
-  // Pack grouping (AI-driven variant clustering)
-  // -----------------------------------------------------------------------
-
-  /** Return the cached pack mapping if it's up-to-date with the current
-   *  library. Returns null when an import is needed. */
-  getPackMapping(): Promise<Record<string, string> | null>;
-  /** Build the prompt text the user should send to Claude to generate
-   *  the pack grouping. The user copies this, pastes it into Claude,
-   *  and imports the JSON response back. */
-  exportPackGroupingPrompt(): Promise<string>;
-  /** Open a file picker for a .json file, parse and cache the pack
-   *  mapping from it. Returns the mapping, or null if the user cancelled. */
-  importPackMappingFromFile(): Promise<Record<string, string> | null>;
-  /** Merge multiple pack names into one and persist the change. */
-  mergePacks(args: { sourcePacks: string[]; targetName: string }): Promise<Record<string, string>>;
-
-  // -----------------------------------------------------------------------
-  // Item browser
-  // -----------------------------------------------------------------------
-
-  /** Search/filter items from the PF2e database. Returns lightweight rows
-   *  without descriptions. Returns [] if the PF2e DB is not configured. */
-  searchItemsBrowser(params: ItemSearchParams): Promise<ItemBrowserRow[]>;
-  /** Full item detail including cleaned description and parsed variants.
-   *  Returns null if the item is not found or the DB is not configured. */
-  getItemBrowserDetail(id: string): Promise<ItemBrowserDetail | null>;
-  /** Distinct filter values (traits, sources, usage categories) for the
-   *  item filter panel. Returns empty facets if the DB is not configured. */
-  getItemFacets(): Promise<ItemFacets>;
-
-  // -----------------------------------------------------------------------
-  // Auto-Wall (wall detection for VTT import)
-  // -----------------------------------------------------------------------
-
-  /** Whether the Auto-Wall binary is configured and available. */
-  autoWallAvailable(): Promise<boolean>;
-  /** Launch Auto-Wall GUI with the given map image pre-loaded. */
-  autoWallLaunch(fileName: string): Promise<void>;
-  /** Check whether a .uvtt file exists for the given map. */
-  autoWallHasUvtt(fileName: string): Promise<boolean>;
-  /** Read wall segments from the .uvtt file as pixel coordinates.
-   *  Returns null if no .uvtt exists. */
-  autoWallGetWalls(fileName: string): Promise<{
-    walls: number[][];
-    width: number;
-    height: number;
-  } | null>;
-  /** Open a file picker to import a .uvtt file for the given map. */
-  autoWallImportUvtt(fileName: string): Promise<boolean>;
-  /** Read the raw .uvtt sidecar JSON for a map. Returns null if no
-   *  sidecar exists. The returned object can be passed directly to
-   *  foundry-mcp's create_scene_from_uvtt tool. */
-  getMapUvtt(fileName: string): Promise<Record<string, unknown> | null>;
-  /** Push a map + its .uvtt walls to Foundry VTT via foundry-mcp.
-   *  Requires foundryMcpUrl to be set in config.json. */
-  pushToFoundry(fileName: string): Promise<{
-    sceneId: string;
-    sceneName: string;
-    wallsCreated: number;
-    doorsCreated: number;
-  }>;
-
-  // -----------------------------------------------------------------------
-  // Monster browser
-  // -----------------------------------------------------------------------
-
-  /** Search/filter monsters from the PF2e database. */
-  monstersSearch(params: MonsterSearchParams): Promise<MonsterSummary[]>;
-  /** Distinct facet values for the filter panel. */
-  monstersFacets(): Promise<MonsterFacets>;
-  /** Full stat block for a single monster by name. */
-  monstersGetDetail(name: string): Promise<MonsterDetail | null>;
-
-  // -----------------------------------------------------------------------
-  // Compendium configuration (Settings → Monsters)
-  // -----------------------------------------------------------------------
-
-  /** List every compendium pack foundry-mcp knows about. Optionally
-   *  narrowed by `documentType` ('Actor' for bestiaries, 'Item' for
-   *  equipment/feats/etc). Used by the Settings dialog to render a live
-   *  multi-select of the packs the user actually has installed. */
-  compendiumListPacks(documentType?: string): Promise<CompendiumPackSummary[]>;
-  /** Currently-active monster pack list — either the user's saved
-   *  override or the hardcoded defaults. */
-  compendiumGetMonsterPackIds(): Promise<string[]>;
-  /** Replace the monster pack list. Passing `[]` resets to defaults.
-   *  Returns the post-save active list so the renderer can update state
-   *  without a follow-up read. */
-  compendiumSetMonsterPackIds(ids: string[]): Promise<string[]>;
-  /** The hardcoded default monster pack list. Exposed so Settings can
-   *  offer a "Reset to default" button that shows the default tick set
-   *  without hard-coding it in the renderer. */
-  compendiumGetDefaultMonsterPackIds(): Promise<string[]>;
-
-  // -----------------------------------------------------------------------
-  // Globe pins
-  // -----------------------------------------------------------------------
-
-  /** All saved globe pins. */
-  globePinsList(): Promise<GlobePin[]>;
-  /** Create or update a globe pin (upsert by id). */
-  globePinsUpsert(pin: GlobePin): Promise<void>;
-  /** Delete a globe pin by id. */
-  globePinsDelete(id: string): Promise<void>;
-  /** Open (or create) the Obsidian note for a globe pin. Returns false if no vault configured. */
-  globePinOpenNote(pin: GlobePin): Promise<boolean>;
-  /** Load the mission briefing data for a mission pin by parsing its Obsidian note frontmatter. */
-  globePinGetMission(pin: GlobePin): Promise<MissionData | null>;
-  /** Associate a pin with an existing Obsidian note via native file picker.
-   *  Stamps the pin id into the note's frontmatter so rename-resilient
-   *  lookup continues to work, then updates the pin's stored note path.
-   *  Returns the updated pin, or null if the user cancelled or the chosen
-   *  file is outside the vault. */
-  globePinLinkNote(pin: GlobePin): Promise<GlobePin | null>;
-  /** Export all pins + parsed mission data to a JSON file via save dialog.
-   *  The output file is designed for the player-map static site. */
-  globeExportPlayerData(): Promise<boolean>;
-  // -----------------------------------------------------------------------
-  // Party inventory (live-synced via foundry-mcp)
-  // -----------------------------------------------------------------------
-
-  inventoryList(): Promise<PartyInventoryItem[]>;
-  inventoryUpsert(item: PartyInventoryItem): Promise<void>;
-  inventoryDelete(id: string): Promise<void>;
-
-  // -----------------------------------------------------------------------
-  // Aurus leaderboard (live-synced via foundry-mcp)
-  // -----------------------------------------------------------------------
-
-  aurusList(): Promise<AurusTeam[]>;
-  aurusUpsert(team: AurusTeam): Promise<void>;
-  aurusDelete(id: string): Promise<void>;
-
-  // -----------------------------------------------------------------------
-  // Combat tracker (encounters + initiative)
-  // -----------------------------------------------------------------------
-
-  encountersList(): Promise<Encounter[]>;
-  encountersUpsert(encounter: Encounter): Promise<void>;
-  encountersDelete(id: string): Promise<void>;
-  /** Subscribe to live initiative updates pushed from Foundry via the
-   *  combat SSE channel. Fires when a combatant's initiative changes in
-   *  Foundry (e.g. a player rolls initiative). Returns an unsubscribe fn. */
-  onCombatantInitiativeUpdate(callback: (event: CombatantInitiativeEvent) => void): () => void;
-  /** Generate loot for an encounter via Anthropic. Returns the new loot
-   *  list; the renderer is responsible for persisting it onto the
-   *  encounter via encountersUpsert. */
-  generateEncounterLoot(args: { encounter: Encounter; partyLevel: number; apiKey: string }): Promise<LootItem[]>;
-  /** Push an encounter's monster combatants to Foundry VTT as actors,
-   *  organized in a folder named after the encounter. Requires
-   *  foundryMcpUrl to be set in config.json and a live Foundry session
-   *  with the API Bridge module connected. Returns a summary of what
-   *  was created + skipped so the UI can surface ambiguities. */
-  pushEncounterToFoundry(encounterId: string): Promise<PushEncounterResult>;
-  /** Fetch player characters from the GM's party folder in Foundry.
-   *  Returns an empty array when foundryMcpUrl is not configured, the
-   *  bridge is not connected, or the folder contains no characters.
-   *  The folder name defaults to "The Party" and can be overridden by
-   *  passing `?folder=` on the HTTP side. */
-  listPartyMembers(): Promise<PartyMember[]>;
-  /** Fetch spellcasting entries + slot state for a Foundry actor by id.
-   *  Returns null when foundryMcpUrl is not configured or the bridge is
-   *  disconnected. */
-  getActorSpellcasting(actorId: string): Promise<ActorSpellcasting | null>;
-  /** Subscribe to live actor state changes from Foundry via the `actors`
-   *  SSE channel. Fires on any `updateActor` hook; `changedPaths` lets
-   *  handlers filter to the fields they care about. Returns an unsubscribe
-   *  function. */
-  onActorUpdated(callback: (update: ActorUpdate) => void): () => void;
-  /** Push a manual HP change from dm-tool back to Foundry. PATCHes
-   *  `system.attributes.hp.value` (and `.max` if provided) on the actor
-   *  via foundry-mcp. No-op when foundryMcpUrl is not configured. */
-  pushActorHp(actorId: string, hp: number, maxHp?: number): Promise<void>;
-}
-
-// --- Party inventory ---------------------------------------------------------
-
-export type PartyInventoryCategory = 'consumable' | 'equipment' | 'quest' | 'treasure' | 'other';
-
-export interface PartyInventoryItem {
-  id: string;
-  name: string;
-  qty: number;
-  category: PartyInventoryCategory;
-  bulk?: number;
-  /** Price of a single unit in copper pieces. Multiply by qty for total value. */
-  valueCp?: number;
-  /** Link to the Archives of Nethys entry, if known. */
-  aonUrl?: string;
-  note?: string;
-  /** Who's carrying this — character name, "Party" for shared, or undefined. */
-  carriedBy?: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-// --- Aurus leaderboard -------------------------------------------------------
-
-export interface AurusTeam {
-  id: string;
-  name: string;
-  /** Free-text emblem descriptor (could become a game-icons name later). */
-  emblem?: string;
-  /** CSS color string for the team banner stripe. */
-  color: string;
-  /** Open-ended combat rating. DM-adjusted; no implicit ceiling. */
-  combatPower: number;
-  /** Total loot recovered, in copper pieces. */
-  valueReclaimedCp: number;
-  /** Exactly one team should be flagged true; the player portal highlights it. */
-  isPlayerParty: boolean;
-  note?: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-// --- Party members (Foundry live query) --------------------------------------
-
-/** A player character fetched from the GM's party folder in Foundry.
- *  Stats are pre-extracted for the combat tracker so the picker
- *  can display them without a follow-up actor fetch. */
-export interface PartyMember {
-  id: string;
-  name: string;
-  img: string;
-  /** Perception modifier (PF2e `system.perception.mod`), used as the
-   *  default initiative modifier in the combat tracker. */
-  initiativeMod: number;
-  /** Current HP at the time the picker fetched the party. The combat
-   *  tracker uses this as the combatant's starting HP so a PC mid-fight
-   *  isn't reset to full when added to the encounter. */
-  hp: number;
-  maxHp: number;
-}
-
-// --- Combat tracker ----------------------------------------------------------
-
-export type LootKind = 'currency' | 'item' | 'consumable' | 'narrative';
-export type LootSource = 'db' | 'ai' | 'manual';
-
-export interface LootItem {
-  id: string;
-  name: string;
-  /** Short description or flavor text — one or two sentences from the AI,
-   *  or a DM-authored note for manual entries. */
-  description: string;
-  kind: LootKind;
-  /** Unit value in copper pieces (multiply by qty for total value). */
-  valueCp?: number;
-  qty: number;
-  /** pf2e-db `items.id` when this row was drawn from the database. Lets the
-   *  "send to inventory" action link back to a known item for stats. */
-  itemId?: string;
-  /** Canonical Archives of Nethys page for the item, when known. */
-  aonUrl?: string;
-  /** Where this row came from — distinguishes AI-invented from DB-drawn
-   *  items in the UI so the DM can audit before committing. */
-  source: LootSource;
-}
-
-export type CombatantKind = 'monster' | 'pc';
-
-export interface Combatant {
-  /** Stable UUID within this encounter — used as React key and for edits. */
-  id: string;
-  kind: CombatantKind;
-  /** Exact monster name from pf2e-db. Only present for kind='monster' — used
-   *  to refetch the full stat block on demand. */
-  monsterName?: string;
-  /** Rendered name. Auto-numbered ("Goblin 1", "Goblin 2") when multiple of
-   *  the same monster are added, but freely editable by the DM. */
-  displayName: string;
-  /** Initiative modifier (Perception for monsters by default). Used for the
-   *  auto-roll button and kept as the tiebreaker when two combatants roll
-   *  the same total. */
-  initiativeMod: number;
-  /** Rolled initiative total. null before the encounter has been rolled —
-   *  unrolled combatants sort to the end of the order. */
-  initiative: number | null;
-  hp: number;
-  maxHp: number;
-  /** Free-form conditions / status notes. */
-  notes?: string;
-  /** Foundry actor document id. Set only when the PC was added from the party
-   *  picker (where we have the live actor id). Absent for manually-entered PCs
-   *  and monsters. Required by the spell cast + slot display features, the
-   *  live HP sync via the `actors` SSE channel, and to match incoming
-   *  `updateCombatant` SSE events so the tracker updates automatically when
-   *  a player rolls initiative in Foundry. */
-  foundryActorId?: string;
-}
-
-/** Payload pushed from the Electron main process to the renderer whenever
- *  a Foundry actor changes via the `actors` SSE channel. `changedPaths` is
- *  the dot-notation diff from the `updateActor` hook; `system` is the full
- *  PF2e `actor.getRollData()` snapshot at the time of the event. Renderer
- *  hooks filter by path and extract the fields they care about. */
-export interface ActorUpdate {
-  actorId: string;
-  changedPaths: string[];
-  system: Record<string, unknown>;
-}
-
-/** Payload pushed over IPC when Foundry fires an updateCombatant hook that
- *  sets a new initiative value. The dm-tool main process subscribes to the
- *  foundry-mcp `combat` SSE channel and forwards these to the renderer. */
-export interface CombatantInitiativeEvent {
-  /** Foundry combat encounter id (for debugging/logging). */
-  encounterId: string;
-  /** Foundry actor id — matches `Combatant.foundryActorId`. */
-  actorId: string;
-  /** The newly-rolled initiative total. */
-  initiative: number;
-}
-
-/** One monster combatant successfully turned into a Foundry actor. */
-export interface PushedActorSummary {
-  displayName: string;
-  monsterName: string;
-  actorId: string;
-  actorName: string;
-  actorUuid: string;
-  sourcePackId: string;
-  sourcePackLabel: string;
-}
-
-/** One combatant we couldn't push — either no compendium hit, no monsterName,
- *  or Foundry rejected the create. */
-export interface SkippedCombatantSummary {
-  displayName: string;
-  monsterName?: string;
-  reason: string;
-}
-
-export interface PushEncounterResult {
-  /** Folder the actors were placed in. null when nothing was pushed. */
-  folderId: string | null;
-  folderName: string | null;
-  /** True when a new folder was created; false when an existing one was
-   *  reused (re-push of the same encounter name). */
-  folderCreated: boolean;
-  created: PushedActorSummary[];
-  skipped: SkippedCombatantSummary[];
-}
-
-export interface Encounter {
-  id: string;
-  name: string;
-  combatants: Combatant[];
-  /** Index into combatants[] (sorted-order, see below) pointing at whose
-   *  turn it currently is. Bounds-checked against combatants.length by the
-   *  UI — persisted as-is. */
-  turnIndex: number;
-  /** 1-indexed round counter, incremented when the turn pointer wraps. */
-  round: number;
-  /** Treasure awarded for this encounter. Populated manually or via the
-   *  AI auto-generate button. */
-  loot: LootItem[];
-  /** When true, the AI is allowed to self-author up to ~20% of the loot;
-   *  the rest must be drawn from pf2e-db items. When false, every row must
-   *  come from the DB. */
-  allowInventedItems: boolean;
-  createdAt: string;
-  updatedAt: string;
-}
-
-// --- Spell cast + slot display (combat panel) --------------------------------
-
-export type SpellPreparationMode = 'prepared' | 'spontaneous' | 'innate' | 'focus' | 'ritual' | 'items';
-
-export interface CombatSpellSummary {
-  id: string;
-  name: string;
-  /** Base spell rank (0 = cantrip). */
-  rank: number;
-  isCantrip: boolean;
-  /** PF2e action cost string: "1" | "2" | "3" | "reaction" | "free" | etc. */
-  actions: string;
-  /** Prepared mode only — true when this prepared slot has been expended today. */
-  expended?: boolean;
-  /** Trait slugs (cantrip excluded). Used for hover card display. */
-  traits: string[];
-  /** Plain-text range string, e.g. "30 feet". Empty string when absent. */
-  range: string;
-  /** Plain-text area string, e.g. "15-foot cone". Empty string when absent. */
-  area: string;
-  /** Plain-text targets string. Empty string when absent. */
-  target: string;
-  /** Plain text description (Foundry markup stripped). May be empty. */
-  description: string;
-}
-
-/** Per-rank slot count for spontaneous casters. */
-export interface CombatSpellSlot {
-  rank: number;
-  value: number;
-  max: number;
-}
-
-export interface CombatSpellEntry {
-  id: string;
-  name: string;
-  mode: SpellPreparationMode;
-  tradition: string;
-  spells: CombatSpellSummary[];
-  /** Spontaneous only — slot state per rank, ranks with max=0 omitted. */
-  slots?: CombatSpellSlot[];
-  /** Focus only — shared focus point pool. */
-  focusPoints?: { value: number; max: number };
-}
-
-export interface ActorSpellcasting {
-  actorId: string;
-  entries: CombatSpellEntry[];
 }
